Gui parts created, along with a map, and removed Tier system in favor of a locking/unlocking class
Created the FXML files for the mainscreen,added an automapping program
and some images, implemented a random land pattern generator. Filled out
a skeleton for a java version of the JSON parser to be used with random
events.. Created DisplayData Package with three classes used to
communicate with the main gui without allowing access to the private
members of the main game code. On update, these data classes will be
sent to the guihandler which will decide how to process them. Instead of
letting the player pick a spot with the map program, I decided to make
it so a world map will be created and directly interacted with by the
player. This required the implementation of a Vector3 class (if ever I
wanted to implement some sort of terrain system).