Time Functions
Added Calendar support and printing; corrected Time's Greater, less than
and equal to function to reflect the addition of the calendar. The
subtraction method broken in most cases and needs fixed to finish
handler classes. Seeing how the correct time function would be broken in
certain cases in the time class, I still need to fix that. Finally, I
decided to make it so that the time class has to be declared and passed
to a Building. This makes it easier than converting it every time.