Pathfinding Fix, NoisyDataHandler, and a FileLogger
Pathfinding seems to be working 100%. I added a FileLogger to view later
on if pathfinding acts weird again. It only writes to the Log directory
when the file is closed. I added an Instant Tooltip thing, but it isn't
working exactly how I want it to work. The Delay is too long. I will
have to override the popup class to get it to work right.