AI Noise Skeleton, a Camp Data 'Struct', Raid Entity
Added the AI Noise Skeleton
This class will be what the AI queries for information. The values will
be salted with a noise of some sort. From that, the AI will examine the
values and attempt to deduce the exact resources the player has.

Added  Camp Data 'Struct'
--Only Skeleton
This is the structure that holds the campdata. Unknown who updates it.
It'll likely be a linked list of sorts so taht it  can average the
values for the AI to try to guess things.

Added RaidEntityAI Skeleton
This controls the RaidEntityObject, calculates the path and responds to
the things it sees. Once activated it can't be recalled unless it sees a
reason to retreat. (Such as a second raid group heading for their home
etc.)

Added Raid Entity Object
This holds who sent the object, it's Path Object that it creates in the
constructor. The current implementation is flawed though.

Worked on the Territory stuff to figure out what tile to use to overlay.
Currently it doesn't take into account if another tribes territory is
close to this one.